+++ app/Imports/Purchases/Sheets/BillItems.php
@@ -24,13 +24,13 @@ class BillItems extends Import
 
         $row['document_id'] = (int) Document::bill()->number($row['bill_number'])->pluck('id')->first();
 
-        if (empty($row['item_id']) && !empty($row['item_name'])) {
+        if (empty($row['item_id']) && ! empty($row['item_name'])) {
             $row['item_id'] = $this->getItemIdFromName($row);
 
             $row['name'] = $row['item_name'];
         }
 
-        $row['tax'] = (double) $row['tax'];
+        $row['tax'] = (float) $row['tax'];
         $row['tax_id'] = 0;
         $row['type'] = Document::BILL_TYPE;
